Fix linker not initialising some global variables
GCC, by default, locates global variables initialised to 0 in bss section.
In practice, this means they were not getting initialised at all;
my theory is our custom linker doesn't contain code to initialise
such variables.

The fix is -fno-zero-initialized-in-bss

Before this, if you had a global (or static var in a function scope,
as these are actually globals) initialised to 0 and tried to read
the value, it would be random noise.